requirements.txt
django
psycopg2
wechat-sdk
qiniu
djangorestframework
ipython
